refactor: 절대경로 추가
- #26
