+++ packages/server/src/product/product.module.ts
@@ -1,9 +1,9 @@
 import { CategoryModule } from '@category/category.module';
 import { forwardRef, Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { LocationModule } from '@src/location/location.module';
-import { UserModule } from '@src/user/user.module';
-import { WishModule } from '@src/wish/wish.module';
+import { LocationModule } from '@location/location.module';
+import { UserModule } from '@user/user.module';
+import { WishModule } from '@wish/wish.module';
 import { Product } from './entities/product.entity';
 import { ProductController } from './product.controller';
 import { ProductService } from './product.service';
