+++ packages/server/src/product/product.service.ts
@@ -2,12 +2,12 @@ import { CategoryService } from '@category/category.service';
 import { PRODUCT_QUERY } from '@constant/queries';
 import { HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { CustomException } from '@src/base/CustomException';
-import { productStatus } from '@src/constant/enum';
-import { ErrorMessage } from '@src/constant/ErrorMessage';
-import { LocationService } from '@src/location/location.service';
-import { UserService } from '@src/user/user.service';
-import { WishService } from '@src/wish/wish.service';
+import { CustomException } from '@base/CustomException';
+import { productStatus } from '@constant/enum';
+import { ErrorMessage } from '@constant/ErrorMessage';
+import { LocationService } from '@location/location.service';
+import { UserService } from '@user/user.service';
+import { WishService } from '@wish/wish.service';
 import { ProductInsertDto } from './dto/productInsert.dto';
 import { ProductSearchDto } from './dto/productSearch.dto';
 import { Product, ProductRepository } from './entities/product.entity';
