+++ packages/server/src/user/user.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ProductModule } from '@product/product.module';
-import { LocationModule } from '@src/location/location.module';
-import { WishModule } from '@src/wish/wish.module';
+import { LocationModule } from '@location/location.module';
+import { WishModule } from '@wish/wish.module';
 import { UserLocationModule } from '@userLocation/userLocation.module';
 import { User } from './entities/user.entity';
 import { UserController } from './user.controller';
